requirements.txt
# Core Flask
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3

# Forms and Validation
Flask-WTF==1.1.1
WTForms==3.0.1
email-validator==2.0.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.7  # PostgreSQL driver

# Security and Cryptography
bcrypt==4.0.1
cryptography==41.0.7
PyJWT==2.8.0  # Para tokens JWT

# Email
Flask-Mail==0.9.1

# File handling and uploads
Werkzeug==2.3.7

# Environment and Configuration
python-dotenv==1.0.0

# Development and Testing
Flask-DebugToolbar==0.13.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8
mkdocs-material-extensions==1.3.1
pymdown-extensions==10.7
Markdown==3.5.1
Pygments==2.18.0
pymdown-extensions==10.16.1

# Production Deployment
gunicorn==21.2.0

# BI & Analytics
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
openpyxl==3.1.2
reportlab==4.0.7
xhtml2pdf==0.2.11

# Optional: Additional utilities
click==8.1.7
itsdangerous==2.1.2
blinker==1.6.2